feat: migrate to app directory
