+++ next.config.js
@@ -13,25 +13,6 @@ const nextConfig = {
   //     'res.cloudinary.com',
   //   ],
   // },
-
-  // SVGR
-  webpack(config) {
-    config.module.rules.push({
-      test: /\.svg$/i,
-      issuer: /\.[jt]sx?$/,
-      use: [
-        {
-          loader: '@svgr/webpack',
-          options: {
-            typescript: true,
-            icon: true,
-          },
-        },
-      ],
-    });
-
-    return config;
-  },
 };
 
 module.exports = nextConfig;
