--- src/__tests__/pages/404.test.tsx
@@ -1,13 +1,0 @@
-import { render, screen } from '@testing-library/react';
-
-import NotFoundPage from '@/pages/404';
-
-describe('404', () => {
-  it('renders a heading', () => {
-    render(<NotFoundPage />);
-
-    const heading = screen.getByText(/not found/i);
-
-    expect(heading).toBeInTheDocument();
-  });
-});
