+++ src/app/api/hello/route.ts
@@ -1,0 +1,5 @@
+import { NextResponse } from 'next/server';
+
+export async function GET() {
+  return NextResponse.json({ hello: 'Next.js' });
+}
