+++ src/app/components/layout.tsx
@@ -1,0 +1,15 @@
+import { Metadata } from 'next';
+import * as React from 'react';
+
+export const metadata: Metadata = {
+  title: 'Components',
+  description: 'Pre-built components with awesome default',
+};
+
+export default function ComponentsLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <>{children}</>;
+}
