--- src/components/layout/Layout.tsx
@@ -1,6 +1,0 @@
-import * as React from 'react';
-
-export default function Layout({ children }: { children: React.ReactNode }) {
-  // Put Header or Footer Here
-  return <>{children}</>;
-}
