+++ src/constant/config.ts
@@ -1,0 +1,7 @@
+export const siteConfig = {
+  title: 'Next.js + Tailwind CSS + TypeScript Starter',
+  description:
+    'A starter for Next.js, Tailwind CSS, and TypeScript with Absolute Import, Seo, Link component, pre-configured with Husky',
+  /** Without additional '/' on the end, e.g. https://theodorusclarence.com */
+  url: 'https://tsnext-tw.thcl.dev',
+};
