+++ src/lib/__tests__/og.test.ts
@@ -1,4 +1,4 @@
-import { openGraph } from '@/lib/helper';
+import { openGraph } from '@/lib/og';
 
 describe('Open Graph function should work correctly', () => {
   it('should not return templateTitle when not specified', () => {
